Fix npc shop static_cast

+++ src/lua/functions/creatures/npc/npc_type_functions.cpp
@@ -193,11 +193,11 @@ int NpcTypeFunctions::luaNpcTypeAddShopItem(lua_State* L) {
 	ShopInfo shopItem;
 
 	shopItem.itemClientId = static_cast<uint16_t>(getField<uint32_t>(L, table, "clientId"));
-	shopItem.buyPrice = static_cast<uint16_t>(getField<uint32_t>(L, table, "buy"));
-	shopItem.sellPrice = static_cast<uint16_t>(getField<uint32_t>(L, table, "sell"));
-	shopItem.subType = static_cast<uint16_t>(getField<uint32_t>(L, table, "count"));
-	shopItem.storageKey = static_cast<uint16_t>(getField<uint32_t>(L, table, "storageKey"));
-	shopItem.storageValue = static_cast<uint16_t>(getField<uint32_t>(L, table, "storageValue"));
+	shopItem.buyPrice = static_cast<uint32_t>(getField<uint32_t>(L, table, "buy"));
+	shopItem.sellPrice = static_cast<uint32_t>(getField<uint32_t>(L, table, "sell"));
+	shopItem.subType = static_cast<int32_t>(getField<uint32_t>(L, table, "count"));
+	shopItem.storageKey = static_cast<int32_t>(getField<uint32_t>(L, table, "storageKey"));
+	shopItem.storageValue = static_cast<int32_t>(getField<uint32_t>(L, table, "storageValue"));
 
 	const ItemType &it = Item::items.getItemIdByClientId(shopItem.itemClientId);
